+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## 3.0.3
+## 3.0.4
  - Jul 18, 2021
  - Docs updated
 
@@ -8,14 +8,14 @@
    - `pages`, `pageBuilders` and `keys` parameters are combined into `pages` parameter. (breaking)
    - User don't need to create keys anymore.
    - Helper methods and documentation added for in-page navigation.
- - Bottom bar delegates are removed.
+ - Bottom bar delegates are removed. (breaking)
    - Bottom navbar is created by the user, using a builder.
    - Removes a layer of adapters in between.
    - Removes all bottom bar dependencies.
  - Parameter defaults are changed to make the user more aware of their design choices.
    - `savePageState` default is change to false.
    - `pageStack` default is change to `NoPageStack` for iOS apps.
- - Page transition animations added.
+ - Page transition animations added.8
    - Use `pageTransitionData` parameter to add animations while changing the page.
  - State management inside the library and performance is improved.
  - Comprehensive documentation added.